+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # mhtml2html Change Log
 
+## [v1.0.3]
+
+** Bug fixes **
+
+- Fix bundling runtime dependencies
+- Fix `html_only` support in mhtml2html parsing for node.js
+
 ## [v1.0.0]
 
 ** Features **